+++ Tests/Unit/Widgets/Provider/GenericDoughnutChartDataProviderTest.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
 namespace Brotkrueml\MatomoWidgets\Tests\Unit\Widgets\Provider;
 
 use Brotkrueml\MatomoWidgets\Connection\ConnectionConfiguration;
-use Brotkrueml\MatomoWidgets\Domain\Repository\MatomoRepositoryInterface;
+use Brotkrueml\MatomoWidgets\Domain\Repository\MatomoRepository;
 use Brotkrueml\MatomoWidgets\Extension;
 use Brotkrueml\MatomoWidgets\Parameter\ParameterBag;
 use Brotkrueml\MatomoWidgets\Widgets\Provider\GenericDoughnutChartDataProvider;
@@ -25,7 +25,7 @@ class GenericDoughnutChartDataProviderTest extends TestCase
     private ConnectionConfiguration $connectionConfiguration;
 
     /**
-     * @var Stub|MatomoRepositoryInterface
+     * @var Stub|MatomoRepository
      */
     private $repositoryStub;
 
@@ -37,7 +37,7 @@ class GenericDoughnutChartDataProviderTest extends TestCase
     protected function setUp(): void
     {
         $this->connectionConfiguration = new ConnectionConfiguration('https://example.org/', 1, '');
-        $this->repositoryStub = $this->createStub(MatomoRepositoryInterface::class);
+        $this->repositoryStub = $this->createStub(MatomoRepository::class);
         $this->languageServiceStub = $this->createStub(LanguageService::class);
     }