+++ Tests/Unit/Widgets/Provider/GenericTableDataProviderTest.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
 namespace Brotkrueml\MatomoWidgets\Tests\Unit\Widgets\Provider;
 
 use Brotkrueml\MatomoWidgets\Connection\ConnectionConfiguration;
-use Brotkrueml\MatomoWidgets\Domain\Repository\MatomoRepositoryInterface;
+use Brotkrueml\MatomoWidgets\Domain\Repository\MatomoRepository;
 use Brotkrueml\MatomoWidgets\Parameter\ParameterBag;
 use Brotkrueml\MatomoWidgets\Widgets\Decorator\DecoratorInterface;
 use Brotkrueml\MatomoWidgets\Widgets\Provider\GenericTableDataProvider;
@@ -26,7 +26,7 @@ class GenericTableDataProviderTest extends TestCase
     private ConnectionConfiguration $connectionConfiguration;
 
     /**
-     * @var MockObject|MatomoRepositoryInterface
+     * @var MockObject|MatomoRepository
      */
     private MockObject $repositoryMock;
 
@@ -38,7 +38,7 @@ class GenericTableDataProviderTest extends TestCase
     protected function setUp(): void
     {
         $this->connectionConfiguration = new ConnectionConfiguration('https://example.org/', 1, '');
-        $this->repositoryMock = $this->createMock(MatomoRepositoryInterface::class);
+        $this->repositoryMock = $this->createMock(MatomoRepository::class);
         $this->languageServiceStub = $this->createStub(LanguageService::class);
     }