+++ Tests/Unit/Widgets/Provider/GenericValueDataProviderTest.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
 namespace Brotkrueml\MatomoWidgets\Tests\Unit\Widgets\Provider;
 
 use Brotkrueml\MatomoWidgets\Connection\ConnectionConfiguration;
-use Brotkrueml\MatomoWidgets\Domain\Repository\MatomoRepositoryInterface;
+use Brotkrueml\MatomoWidgets\Domain\Repository\MatomoRepository;
 use Brotkrueml\MatomoWidgets\Parameter\ParameterBag;
 use Brotkrueml\MatomoWidgets\Widgets\Provider\GenericValueDataProvider;
 use PHPUnit\Framework\MockObject\Stub;
@@ -22,7 +22,7 @@ class GenericValueDataProviderTest extends TestCase
 {
     private ConnectionConfiguration $connectionConfiguration;
     /**
-     * @var Stub|MatomoRepositoryInterface
+     * @var Stub|MatomoRepository
      */
     private $repositoryStub;
     private GenericValueDataProvider $subject;
@@ -30,7 +30,7 @@ class GenericValueDataProviderTest extends TestCase
     protected function setUp(): void
     {
         $this->connectionConfiguration = new ConnectionConfiguration('https://example.org/', 1, '');
-        $this->repositoryStub = $this->createStub(MatomoRepositoryInterface::class);
+        $this->repositoryStub = $this->createStub(MatomoRepository::class);
         $this->subject = new GenericValueDataProvider(
             $this->repositoryStub,
             $this->connectionConfiguration,