+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Configuration independent of a website (#41)
+- PSR-14 event BeforeMatomoApiRequestEvent to adjust site ID and auth token (#42)
 
 ## [1.6.0] - 2022-09-19